chore: add app.js and move previous content from index.js
